+++ test/unit/fuzz-tools.test.js
@@ -86,6 +86,7 @@ import * as billingV81 from '../../lib/billing-v81.js';
 import * as billingV82 from '../../lib/billing-v82.js';
 import * as billingV83 from '../../lib/billing-v83.js';
 import * as toxV86 from '../../lib/tox-v86.js';
+import * as hemodynamicsV87 from '../../lib/hemodynamics-v87.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -116,6 +117,7 @@ const MODULES = {
   'billing-v82.js': billingV82,
   'billing-v83.js': billingV83,
   'tox-v86.js': toxV86,
+  'hemodynamics-v87.js': hemodynamicsV87,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];